Moved password strength component to Auth.
