+++ libs/angular/src/auth/components/password-strength.component.ts
